ver final

+++ include/Fecho.hpp
@@ -1,6 +1,7 @@
 #ifndef FECHO_H
 #define FECHO_H
 
+#include <cmath>
 #include "Ponto.hpp"
 
 class Fecho{
@@ -10,10 +11,10 @@ class Fecho{
 
         Fecho(Ponto* _pontos);
 
-        Ponto* JarvisFecho(Ponto _pontos[], long int _n, long int& _tamanhoFecho);
-        Ponto* MergeConvexHullGraham(Ponto _pontos[], long int _n, long int& _tamanhoFecho);
-        Ponto* InsertConvexHullGraham(Ponto _pontos[], long int _n, long int& _tamanhoFecho);
-        Ponto* BucketConvexHullGraham(Ponto _pontos[], long int _n, long int& _tamanhoFecho);
+        Ponto* JarvisFecho(Ponto _pontos[], int _n, int& _tamanhoFecho);
+        Ponto* MergeConvexHullGraham(Ponto _pontos[], int _n, int& _tamanhoFecho);
+        Ponto* InsertConvexHullGraham(Ponto _pontos[], int _n, int& _tamanhoFecho);
+        Ponto* BucketConvexHullGraham(Ponto _pontos[], int _n, int& _tamanhoFecho);
 
 };
